chore: move password init stuff to form_validation to support mt5 password inputs
